Improve update process

+++ plugin_info/install.php
@@ -17,48 +17,31 @@
  */
 
 function openzwave_update() {
-	log::add('openzwave', 'error', __('Après toute installation/mise à jour pensez bien à mettre à jour les dépendances Openzwave (voir documentation)', __FILE__));
 	if (!file_exists(dirname(__FILE__) . '/../data')) {
 		mkdir(dirname(__FILE__) . '/../data');
+		if (file_exists('/opt/python-openzwave/xml_backups')) {
+			shell_exec('cp -R /opt/python-openzwave/xml_backups ' . dirname(__FILE__) . '/../data');
+			shell_exec('cp -R /opt/python-openzwave/zwcfg_*.xml ' . dirname(__FILE__) . '/../data');
+			shell_exec('rm -rf /opt/python-openzwave/xml_backups');
+			shell_exec('rm -rf /opt/python-openzwave/zwcfg_*.xml');
+		}
 	}
-	shell_exec('cp -R /opt/python-openzwave/xml_backups ' . dirname(__FILE__) . '/../data');
-	shell_exec('cp -R /opt/python-openzwave/zwcfg_*.xml ' . dirname(__FILE__) . '/../data');
-	shell_exec('rm -rf /opt/python-openzwave/xml_backups');
-	shell_exec('rm -rf /opt/python-openzwave/zwcfg_*.xml');
-	config::save('allowStartDeamon', 0, 'openzwave');
-	echo 'Stop zwave network...';
-	openzwave::stopDeamon();
-	echo "OK\n";
-	echo 'Stop cron...';
 	$cron = cron::byClassAndFunction('openzwave', 'pull');
 	if (is_object($cron)) {
 		$cron->remove();
 	}
-	echo "OK\n";
 	if (count(eqLogic::byType('zwave')) > 0) {
 		log::add('openzwave', 'error', 'Attention vous etes sur la nouvelle version d\'openzwave, des actions de votre part sont necessaire merci d\'aller voir https://jeedom.fr/blog/?p=1576');
 	}
-	if (method_exists('openzwave', 'dependancy_info')) {
-		$dependancy_info = openzwave::dependancy_info();
-		if (isset($dependancy_info['state']) && $dependancy_info['state'] != 'ok') {
-			echo __('Lancement de la mise à jour du serveur openzwave...', __FILE__);
-			openzwave::dependancy_install();
-			echo "OK\n";
-		} else {
-			openzwave::runDeamon();
-		}
-	} else {
+	if (openzwave::deamonRunning()) {
 		openzwave::runDeamon();
 	}
 }
 
 function openzwave_remove() {
+	config::save('allowStartDeamon', 0, 'openzwave');
 	if (openzwave::deamonRunning()) {
 		openzwave::stopDeamon();
-	}
-	$cron = cron::byClassAndFunction('openzwave', 'pull');
-	if (is_object($cron)) {
-		$cron->remove();
 	}
 }
